Consolidate formatBlogDate into lib/blog.ts
Per review, drop the standalone lib/blogDate.ts and move formatBlogDate
back to lib/blog.ts. Moves getAllPosts (the only fs-using piece, used
solely in this page's getStaticProps) inline into pages/about/blog/index.tsx
so lib/blog.ts stays free of node-only imports and is safe to bundle into
the client component that calls formatBlogDate.
