+++ src/lib/blog.ts
@@ -1,14 +1,4 @@
-import fs from "fs";
-import path from "path";
 import { parseISO } from "date-fns";
-
-const postsDirectory = path.join(
-  process.cwd(),
-  "src",
-  "pages",
-  "about",
-  "blog",
-);
 
 export interface BlogPost {
   title: string;
@@ -22,23 +12,13 @@ export interface BlogPostWithSlug extends BlogPost {
   slug: string;
 }
 
-export async function getAllPosts(): Promise<BlogPostWithSlug[]> {
-  const items = fs.readdirSync(postsDirectory, { withFileTypes: true });
-  const slugs = items
-    .filter((item) => item.isDirectory())
-    .map((item) => item.name);
-  const posts = await Promise.all(
-    slugs.map(async (slug) => {
-      const { meta } = (await import(
-        `../pages/about/blog/${slug}/index.mdx`
-      )) as { meta: BlogPost };
-      return { slug, ...meta };
-    }),
-  );
+const BLOG_DATE_FORMATTER = new Intl.DateTimeFormat("en-US", {
+  year: "numeric",
+  month: "long",
+  day: "numeric",
+  timeZone: "America/Chicago",
+});
 
-  const sortedPosts = posts.toSorted(
-    (a, b) => parseISO(b.date).getTime() - parseISO(a.date).getTime(),
-  );
-
-  return sortedPosts;
+export function formatBlogDate(date: string): string {
+  return BLOG_DATE_FORMATTER.format(parseISO(date));
 }
